+++ src/main/java/com/springstudy/myspringstudy/domain/post/Post.java
@@ -1,4 +1,4 @@
-package com.springstudy.myspringstudy.domain;
+package com.springstudy.myspringstudy.domain.post;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
