--- src/main/java/com/springstudy/myspringstudy/repository/CustomPostRepository.java
@@ -1,10 +1,0 @@
-package com.springstudy.myspringstudy.repository;
-
-import com.springstudy.myspringstudy.domain.Post;
-import com.springstudy.myspringstudy.dto.request.PostSearch;
-
-import java.util.List;
-
-public interface CustomPostRepository {
-    List<Post> getList(PostSearch postSearch);
-}
