+++ src/main/java/com/springstudy/myspringstudy/repository/post/CustomPostRepository.java
@@ -1,0 +1,10 @@
+package com.springstudy.myspringstudy.repository.post;
+
+import com.springstudy.myspringstudy.domain.post.Post;
+import com.springstudy.myspringstudy.dto.post.request.PostSearch;
+
+import java.util.List;
+
+public interface CustomPostRepository {
+    List<Post> getList(PostSearch postSearch);
+}
