+++ src/main/java/com/springstudy/myspringstudy/repository/post/PostRepository.java
@@ -1,7 +1,7 @@
-package com.springstudy.myspringstudy.repository;
+package com.springstudy.myspringstudy.repository.post;
 
 
-import com.springstudy.myspringstudy.domain.Post;
+import com.springstudy.myspringstudy.domain.post.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long>, CustomPostRepository {
