+++ src/main/java/com/springstudy/myspringstudy/service/post/PostService.java
@@ -1,12 +1,12 @@
-package com.springstudy.myspringstudy.service;
+package com.springstudy.myspringstudy.service.post;
 
-import com.springstudy.myspringstudy.domain.Post;
-import com.springstudy.myspringstudy.dto.request.PostCreate;
-import com.springstudy.myspringstudy.dto.request.PostSearch;
-import com.springstudy.myspringstudy.dto.request.PostUpdate;
-import com.springstudy.myspringstudy.dto.response.PostResponse;
-import com.springstudy.myspringstudy.exception.PostNotFound;
-import com.springstudy.myspringstudy.repository.PostRepository;
+import com.springstudy.myspringstudy.domain.post.Post;
+import com.springstudy.myspringstudy.dto.post.request.PostCreate;
+import com.springstudy.myspringstudy.dto.post.request.PostSearch;
+import com.springstudy.myspringstudy.dto.post.request.PostUpdate;
+import com.springstudy.myspringstudy.dto.post.response.PostResponse;
+import com.springstudy.myspringstudy.exception.post.PostNotFound;
+import com.springstudy.myspringstudy.repository.post.PostRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
